Rework Monument Activator
Refactor activator to take in a toggle action and track the index and activator. This way we can define entity-specific behavior within the entity instead of hardcoding it in the component.

Refactored some of pressure plate as well, since behavior was pretty confusing with static mover vs. button vs. plate toggling.

Also renamed files :^)